ble_fota: Adding 2 steps before starting a firmware update
The init function must be called before the start function.
Moreover, the McuMgrTransport can now implements McuMgrMtuProvider.
If so, the UpdateManager will call the fetchMtu function, and will
feed the new mtu to the ImageManager.

This has a side effect: the setUploadMtu function needs the image
data to check if the new mtu is correct. However, the image data
is only given when the start() function is called. To solve this
issue, the ImageManager now accepts a new constructor with the
image data.

Signed-off-by: Giuliano Franchetto <giuliano.franchetto@intellinium.com>