ble_fota: Rework of the McuMgr response + TLV implementation
The HASH calculus was wrong. A TLV parser has been implemented to find the
hash hidden in the bin file.
This parser can then be modified when the embedded part will be updated.

The response are now directly deserialized without the use of a manager.
This workflow is cleaner and more direct.

Signed-off-by: Giuliano Franchetto <giuliano.franchetto@intellinium.com>