Reimplement BLE transport using Nordics BleManager
This is a reimplementation of the original BLE transporter which uses
Nordic's BleManager to manage the BLE state machine. The purpose of this
reimplementation was twofold: First, the initial implementation gave
control of opening and closing the GATT client to the McuMgr repo which
not ideal. Second, using the Nordic BleManger gives the developer the
option to queue actions to the transporter without having to close the
connection or change ownership of the BluetoothGatt object.

Since Nordic's BleManager does not have a maven repository, I decided to
just paste the code into this project without modifying anything other than
the packages an imports to get it compiling. This is not ideal, but
should be sufficient for now.