Check return code, fail on error after dfu state (#6)
* Check return code, fail on error after dfu state

The firmware upgrade process coud fail in the response handler after most
states due accessing the length of a null array. The values of the
response may be null if the device response with an error return code.

This commit adds a check on the return code of the response before
accessing the values. If the return code indicates an error on the
device, the firmware upgrade will be failed with an
McuMgrErrorException.