[NO MERGE] Prepare mvn central (#27)
Prepare mcumgr-android for publishing on maven central. 

This commit modifies the gradle build files for hosting on the maven
central repository. The gradle file which is used to upload the archives to
maven central is taken from https://github.com/JakeWharton/timber.

The group details for the project are located in the root
build.properties file. This file contains, among others:
    * Version: 0.6.3-SNAPSHOT
    * Group: io.runtime.mcumgr

The package specific artifact ID and name are located in the module
gradle.properties file. The core and transport specific implementations
have been split. The core now resides in the mcumgr-core module, 
while the BLE transport implementation has a separate module, 
mcumgr-ble, which pulls mcumgr-core in as an api dependency.

This change is useful to allow users to reimplement their own transport
layer or potentially add a new transport all together. Furthermore,
since I plan on adding the CoAP BLE transport at some point to the main
repo, this allows me to simply add a new module, mcumgr-coap-ble, rather
than adding it to mcumgr-android and forceing users to add dependencies
which they will not use.

Finally, I can forsee a future where the core could be broken out into
it's own plain java library for use in non-android applications. This
would simplify that process as it is already built as a separate
artifact.